remove actions-overflow
