+++ src/main/js/components/header/index.js
@@ -1,9 +1,10 @@
-import computeActions from "@/components/header/actions-overflow";
+import updateActionsForTouch from "@/components/header/actions-touch";
 import computeBreadcrumbs from "@/components/header/breadcrumbs-overflow";
 
 function init() {
   // Recompute what actions and breadcrumbs should be visible when the viewport size is changed
   computeOverflow();
+  updateActionsForTouch();
   let lastWidth = window.innerWidth;
   window.addEventListener("resize", () => {
     if (window.innerWidth !== lastWidth) {
@@ -59,7 +60,6 @@ function init() {
 }
 
 function computeOverflow() {
-  computeActions();
   computeBreadcrumbs();
 }
